saving junk code
tried and failed to autodetect where the diary starts.
due to line breaks with Missions, I'm not so sure it can
even be done without a config file that knows all mission
names.

So... gonna change lanes entirely, and manually go through
my Diary to add 'Diary:' or some such tag to make it obvious
to the script where the juicy text starts